// devui/src/app/modules/dnb/models/constants/dnb-routes.constants.ts
export enum DnBRoutes {
  selectDrug = "/dnb/select-drug",
  drugVersions = "/dnb/drug-versions",
  approvedVersion = "/dnb/approved-version",
  newDrug = "/dnb/new-drug",
  newVersion = "/dnb/new-version",
  compareVersions = "/dnb/compare-versions",
  drugsInApprovalProcess = "/dnb/drugs-in-approval-process",
  webCrawling = "/dnb/web-crawling/drug-search/",
}