// devui/src/app/modules/dnb/models/constants/storage.constants.ts
export enum storageCompareCodes {
  drugVersionCompareNameOne = "DNB_SELECTED_DRUG_VERSION_NAME_ONE",
  drugVersionCompareNameTwo = "DNB_SELECTED_DRUG_VERSION_NAME_TWO",
  drugVersionCompareIdOne = "DNB_SELECTED_DRUG_VERSION_ID_ONE",
  drugVersionCompareIdTwo = "DNB_SELECTED_DRUG_VERSION_ID_TWO",
  drugVersionCompareStatusTwo = "DNB_SELECTED_DRUG_VERSION_STATUS_TWO",
  drugVersionCompareStatusOne = "DNB_SELECTED_DRUG_VERSION_STATUS_ONE",
  drugVersionCompareDateOne = "DNB_SELECTED_DRUG_VERSION_DATE_ONE",
  drugVersionCompareDateTwo = "DNB_SELECTED_DRUG_VERSION_DATE_TWO",
}
export enum storageDrug {
  drugName = "DNB_SELECTED_DRUG_NAME",
  drugLatestVersion = "DNB_SELECTED_DRUG_LATEST_VERSION",
  drugCode = "DNB_SELECTED_DRUG_CODE",
  drugVersion = "DNB_SELECTED_DRUG_VERSION",
  drugDate = "DNB_SELECTED_DRUG_DATE",
  approvedDrugVersion = "DNB_APPROVED_SELECTED_DRUG_VERSION",
  majorVersion = "DNB_SELECTED_MAJOR_VERSION",
  filterSelectDrugABC = "DNB_SELECTED_ABC_FILTER",
  filterSelectDrug = "DNB_SELECTED_DRUG_FILTER",
  filterSelectDrugList = "DNB_SELECTED_DRUG_LIST_FILTER",
  filterSelectDrugApproval = "DNB_SELECTED_DRUG_APPROVAL_FILTER",
  filterMyWorkEditor = "DNB_SELECTED_MY_WORK_EDITOR_FILTER",
  filterMyWorkApprover = "DNB_SELECTED_MY_WORK_APPROVER_FILTER",
  filterMyWork = "DNB_SELECTED_MY_WORK_FILTER",
  filterInReview = "DNB_SELECTED_IN_REVIEW_FILTER",
  filterInProgress = "DNB_SELECTED_IN_PROGRESS_FILTER",
  filterInSubmitted = "DNB_SELECTED_IN_SUBMITTED_FILTER",
  copySectionSource = "DNB_PASTE_SECTION_SOURCE",
  newVersionEditingMode = "DNB_NEW_VERSION_EDITING_MODE",
  newDrugEditingMode = "DNB_NEW_DRUG_EDITING_MODE",
  midRulesAdmin = "DNB_MID_RULE_ADMIN",
  childSectionAutopopulate = "DNB_CHILD_AUTOPOPULATE_SECTION",
  childGlobalReviewSectionAutopopulate = "DNB_CHILD_GLOBAL_REVIEW_AUTOPOPULATE_SECTION",
  childGlobalReviewIcd10CodesSectionAutopopulate = "DNB_CHILD_GLOBAL_REVIEW_ICD10_CODES_AUTOPOPULATE_SECTION",
}

export enum storageCopy {
  copyColumn = "DNB_COPY_COLUMN_VALUE",
}

export enum storageGeneral {
  isAutosaveActive = "isAutosaveActive",
  userSession = "userSession",
  dnbPermissions = "dnbPermissions",
  dnbVersionColumns = "dnbVersionColumns",
  breadcrumsStorage = "breadcrumsStorage",
}